Remove mui dependency
