+++ package.json
@@ -15,8 +15,6 @@
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@fontsource/poppins": "^4.5.1",
-    "@mui/lab": "^5.0.0-alpha.64",
-    "@mui/material": "^5.2.8",
     "ethers": "^5.5.3",
     "framer-motion": "^5.6.0",
     "next": "12.0.7",
