+++ src/components/WalletConnectButton/index.tsx
@@ -1,7 +1,8 @@
-import { useState, useCallback } from 'react';
-import LoadingButton from '@mui/lab/LoadingButton';
 import { useWeb3 } from '@/context/web3Context';
 import { formatAddress } from '@/utils/helper';
+import theme from '@/utils/theme';
+import { Button, ChakraProvider } from '@chakra-ui/react';
+import { useCallback, useState } from 'react';
 import styles from './index.module.css';
 
 export const WalletConnectButton: React.FC = () => {
@@ -15,26 +16,25 @@ export const WalletConnectButton: React.FC = () => {
   }, [connectWallet]);
 
   return (
-    <div className={styles.container}>
-      {!address ? (
-        <LoadingButton
-          loading={loading}
-          onClick={connect}
-          className={styles.connectWalletButton}
-          sx={{
-            '& .MuiLoadingButton-loadingIndicator': {
-              color: '#fff',
-            },
-          }}
-        >
-          Connect Wallet
-        </LoadingButton>
-      ) : (
-        <div className={styles.userAddress}>
-          Your Address: {ens || formatAddress(address)}
-        </div>
-      )}
-    </div>
+    <ChakraProvider theme={theme}>
+      <div className={styles.container}>
+        {!address ? (
+          <Button
+            loading={loading}
+            onClick={connect}
+            disabled={loading}
+            bgColor='black'
+            textColor='gray.400'
+            _hover={{textColor: 'white'}}
+          >
+            {loading ? 'Connecting Wallet' : 'Connect Wallet' }
+          </Button>
+        ) : (
+          <div className={styles.userAddress}>
+            Your Address: {ens || formatAddress(address)}
+          </div>
+        )}
+      </div></ChakraProvider>
   );
 };
