+++ src/pages/index.tsx
@@ -1,340 +1,128 @@
-import type { NextPage } from 'next';
-import styles from './index.module.css';
-import { WalletConnectButton } from '@/components';
-import { useEffect, useState } from 'react';
-
-import Snackbar from '@mui/material/Snackbar';
-import MuiAlert from '@mui/material/Alert';
-import Avatar from '@mui/material/Avatar';
-import LoadingButton from '@mui/lab/LoadingButton';
-import TextField from '@mui/material/TextField';
-
+import ConnectionsTable from '@/components/ConnectionsTable';
+import { isValidAddr } from '@/utils/helper';
 import { followListInfoQuery, searchUserInfoQuery } from '@/utils/query';
-import { FollowListInfoResp, SearchUserInfoResp, Network } from '@/utils/types';
-import { formatAddress, removeDuplicate, isValidAddr } from '@/utils/helper';
-import { useWeb3 } from '@/context/web3Context';
-import { NAME_SPACE, NETWORK, FIRST } from '@/utils/settings'
-
-const Home: NextPage = () => {
-  const { address, cyberConnect } = useWeb3();
-  const [snackbarOpen, setSnackbarOpen] = useState<boolean>(false);
-  const [snackbarText, setSnackbarText] = useState<string>('');
-
-  const [searchInput, setSearchInput] = useState<string>('');
-  const [searchAddrInfo, setSearchAddrInfo] =
-    useState<SearchUserInfoResp | null>(null);
-  const [searchLoading, setSearchLoading] = useState<boolean>(false);
-
-  const [followListInfo, setFollowListInfo] =
-    useState<FollowListInfoResp | null>(null);
-  const [followLoading, setFollowLoading] = useState<boolean>(false);
-
-  const fetchSearchAddrInfo = async (toAddr: string) => {
-    const resp = await searchUserInfoQuery({
-      fromAddr: address,
-      toAddr,
-      namespace: NAME_SPACE,
-      network: NETWORK,
-    });
-    if (resp) {
-      setSearchAddrInfo(resp);
-    }
-  };
-
-  const handleFollow = async () => {
-    if (!cyberConnect || !searchAddrInfo) {
-      return;
-    }
-
-    try {
-      setFollowLoading(true);
-
-      // Execute connect if the current user is not following the search addrress.
-      if (!searchAddrInfo.followStatus.isFollowing) {
-        await cyberConnect.connect(searchInput);
-
-        // Overwrite the local status of isFollowing
-        setSearchAddrInfo((prev) => {
-          return !!prev
-            ? {
-                ...prev,
-                followStatus: {
-                  ...prev.followStatus,
-                  isFollowing: true,
-                },
-              }
-            : prev;
-        });
-
-        // Add the new following to the current user followings list
-        setFollowListInfo((prev) => {
-          return !!prev
-            ? {
-                ...prev,
-                followingCount: prev.followingCount + 1,
-                followings: {
-                  ...prev.followings,
-                  list: removeDuplicate(
-                    prev.followings.list.concat([searchAddrInfo.identity])
-                  ),
-                },
-              }
-            : prev;
-        });
-
-        setSnackbarText('Follow Success!');
-      } else {
-        await cyberConnect.disconnect(searchInput);
-
-        setSearchAddrInfo((prev) => {
-          return !!prev
-            ? {
-                ...prev,
-                followStatus: {
-                  ...prev.followStatus,
-                  isFollowing: false,
-                },
-              }
-            : prev;
-        });
-
-        setFollowListInfo((prev) => {
-          return !!prev
-            ? {
-                ...prev,
-                followingCount: prev.followingCount - 1,
-                followings: {
-                  ...prev.followings,
-                  list: prev.followings.list.filter((user) => {
-                    return user.address !== searchAddrInfo.identity.address;
-                  }),
-                },
-              }
-            : prev;
-        });
-
-        setSnackbarText('Unfollow Success!');
-      }
-
-      setSnackbarOpen(true);
-    } catch (e) {
-      console.error(e);
-    } finally {
-      setFollowLoading(false);
-    }
-  };
-
-  const handleInputChange = async (value: string) => {
-    setSearchInput(value);
-
-    if (isValidAddr(value) && address && address === searchInput) {
-      setSearchLoading(true);
-      await fetchSearchAddrInfo(value);
-    }
-    setSearchLoading(false);
-  };
+import { FIRST, NAME_SPACE, NETWORK } from '@/utils/settings';
+import theme from '@/utils/theme';
+import { ConnectionsData, FollowListInfoResp, SearchUserInfoResp } from '@/utils/types';
+import { Box, ChakraProvider, Flex, Heading, Input } from '@chakra-ui/react';
+import { useEffect, useRef, useState } from 'react';
+import ConnectionsGraph from '../components/ConnectionsGraph';
+import { connections_data } from "../context/connections.data";
 
 
-  const fetchMore = async (type: 'followings' | 'followers') => {
-    if (!address || !followListInfo) {
-      return;
-    }
+const ConnectionsPage = () => {
+    const cd = Object.entries(connections_data);
+    const [address, setAddress] = useState<string>("0x0c493e5fb71428ba99edcb1bbccd925fdd1f48e0");
 
-    const params =
-      type === 'followers'
-        ? {
-            address,
+    const [searchAddrInfo, setSearchAddrInfo] = useState<SearchUserInfoResp | null>(null);
+    const fetchSearchAddrInfo = async (toAddr: string) => {
+        const resp = await searchUserInfoQuery({
+            fromAddr: address,
+            toAddr,
             namespace: NAME_SPACE,
             network: NETWORK,
-            followerFirst: FIRST,
-            followerAfter: followListInfo.followers.pageInfo.endCursor,
-          }
-        : {
+        });
+        if (resp) {
+            setSearchAddrInfo(resp);
+        }
+    };
+
+    const [followListInfo, setFollowListInfo] =
+        useState<FollowListInfoResp | null>(null);
+
+    const [connections, setConnections] = useState<ConnectionsData>({data: []});
+    useEffect(() => {
+        // TODO: it should be done querying the info for each follower because if there are more than query maximum (FIRST=1000) it could be missing in the list
+        // combine followers and following into connections
+        let connections = followListInfo?.followers.list.map((follower) => ({...follower, is_follower: true, is_following: false}));
+        const following = followListInfo?.followings.list.map((following) => ({...following, is_follower:false, is_following: true}));
+        following?.forEach((fing) => {
+            let connection_index = connections?.findIndex((fer) => {
+                return fer.address == fing.address;
+            });
+            if (connection_index === -1 || connection_index === undefined) {
+                connections?.push(fing);
+            } else {
+                if(connections == undefined) {
+                    connections = [];
+                }
+                connections[connection_index].is_following = true;
+            }
+        });
+        if(connections !== undefined) {
+            setConnections({ data: connections });
+        }
+    }, [followListInfo]);
+
+    useEffect(() => {
+        // Get the current user followings and followers list
+        const initFollowListInfo = async () => {
+            if (!address) {
+                return;
+            }
+
+        const resp = await followListInfoQuery({
             address,
             namespace: NAME_SPACE,
             network: NETWORK,
             followingFirst: FIRST,
-            followingAfter: followListInfo.followings.pageInfo.endCursor,
-          };
-
-    const resp = await followListInfoQuery(params);
-    if (resp) {
-      type === 'followers'
-        ? setFollowListInfo({
-            ...followListInfo,
-            followers: {
-              pageInfo: resp.followers.pageInfo,
-              list: removeDuplicate(
-                followListInfo.followers.list.concat(resp.followers.list)
-              ),
-            },
-          })
-        : setFollowListInfo({
-            ...followListInfo,
-            followings: {
-              pageInfo: resp.followings.pageInfo,
-              list: removeDuplicate(
-                followListInfo.followings.list.concat(resp.followings.list)
-              ),
-            },
-          });
-    }
-  };
-
-  useEffect(() => {
-    // Get the current user followings and followers list
-    const initFollowListInfo = async () => {
-      if (!address) {
-        return;
-      }
-
-      const resp = await followListInfoQuery({
-        address,
-        namespace: NAME_SPACE,
-        network: NETWORK,
-        followingFirst: FIRST,
-        followerFirst: FIRST,
-      });
-      if (resp) {
-        setFollowListInfo(resp);
-      }
+            followerFirst: FIRST,
+        });
+        if (resp) {
+            setFollowListInfo(resp);
+        }
     };
-    initFollowListInfo();
-  }, [address]);
 
-  return (
-    <div className={styles.container}>
-      <div className={styles.logo}>
-        <img
-          src="/cyberconnect-logo.png"
-          alt="CyberConnect Logo"
-          width="100%"
-          height="100%"
-        />
-      </div>
-      <div className={styles.discription}>
-        <p>
-          This is a{' '}
-          <a
-            className={styles.link}
-            href="https://docs.cyberconnect.me/"
-            target="_blank"
-            rel="noreferrer"
-          >
-            CyberConnect
-          </a>{' '}
-          starter app. You can freely use it as a base for your application.{' '}
-        </p>
-        <p>
-          This app displays the current user&apos;s followings and followers. It
-          also allows the user to follow/unfollow a wallet address.
-        </p>
-        <p>Try it yourself!</p>
-      </div>
-      <WalletConnectButton />
-      {address && (
-        <div className={styles.searchSection}>
-          <div className={styles.inputContainer}>
-            <TextField
-              onChange={(e) => handleInputChange(e.target.value)}
-              className={styles.textField}
-              placeholder="Please input the Address you want to follow."
-            />
-            <LoadingButton
-              onClick={handleFollow}
-              disabled={
-                searchLoading ||
-                !isValidAddr(searchInput) ||
-                !address ||
-                address === searchInput
-              }
-              loading={followLoading}
-              className={styles.loadingButton}
-            >
-              {!searchAddrInfo?.followStatus.isFollowing
-                ? 'Follow'
-                : 'Unfollow'}
-            </LoadingButton>
-          </div>
-          {!isValidAddr(searchInput) ? (
-            <div className={styles.error}>Please enter a valid address.</div>
-          ) : address === searchInput ? (
-            <div className={styles.error}>You can’t follow yourself : )</div>
-          ) : (
-            <div className={styles.isFollowed}>
-              This user{' '}
-              {searchAddrInfo?.followStatus.isFollowed
-                ? 'is following you'
-                : 'has not followed you yet'}
-            </div>
-          )}
-        </div>
-      )}
-      {followListInfo && (
-        <div className={styles.listsContainer}>
-          <div className={styles.list}>
-            <div className={styles.subtitle}>
-              You have <strong>{followListInfo.followerCount}</strong>{' '}
-              followers:
-            </div>
-            <div className={styles.followList}>
-              {followListInfo.followers.list.map((user) => {
-                return (
-                  <div key={user.address} className={styles.user}>
-                    <Avatar src={user.avatar} className={styles.avatar} />
-                    <div className={styles.userAddress}>
-                      {user.ens || formatAddress(user.address)}
-                    </div>
-                  </div>
-                );
-              })}
-              {followListInfo.followers.pageInfo.hasNextPage && (
-                <LoadingButton onClick={() => fetchMore('followers')}>
-                  See More
-                </LoadingButton>
-              )}
-            </div>
-          </div>
-          <div className={styles.list}>
-            <div className={styles.subtitle}>
-              You have <strong>{followListInfo.followingCount}</strong>{' '}
-              followings:
-            </div>
-            <div className={styles.followList}>
-              {followListInfo.followings.list.map((user) => {
-                return (
-                  <div key={user.address} className={styles.user}>
-                    <Avatar src={user.avatar} className={styles.avatar} />
-                    <div className={styles.userAddress}>
-                      {user.ens || formatAddress(user.address)}
-                    </div>
-                  </div>
-                );
-              })}
-              {followListInfo.followings.pageInfo.hasNextPage && (
-                <LoadingButton onClick={() => fetchMore('followings')}>
-                  See More
-                </LoadingButton>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-      <Snackbar
-        open={snackbarOpen}
-        autoHideDuration={6000}
-        onClose={() => setSnackbarOpen(false)}
-      >
-        <MuiAlert
-          onClose={() => setSnackbarOpen(false)}
-          severity="success"
-          sx={{ width: '100%' }}
-        >
-          {snackbarText}
-        </MuiAlert>
-      </Snackbar>
-    </div>
-  );
-};
+        initFollowListInfo();
+    }, [address]);
 
-export default Home;
+    const handleInputChange = async (value: string) => {
+        setAddress(value);
+
+        if (isValidAddr(value) && address) {
+            // setSearchLoading(true);
+            await fetchSearchAddrInfo(value);
+        }
+    };
+
+    const [height, setHeight] = useState(200);
+    const [width, setWidth] = useState(200);
+    const graphRef = useRef<HTMLDivElement | null>(null);
+    const onResize = useEffect(() => {
+        if (graphRef !== null && graphRef.current !== null) {
+            setHeight(graphRef.current.getBoundingClientRect().height);
+            setWidth(graphRef.current.getBoundingClientRect().width);
+        }
+        window.onresize = () => {
+            if(graphRef !== null && graphRef.current !== null) {
+                setHeight(graphRef.current.getBoundingClientRect().height);
+                setWidth(graphRef.current.getBoundingClientRect().width);
+            }
+        };
+    },[setWidth, setHeight, graphRef])
+
+    return (
+        <ChakraProvider theme={theme}>
+            <Box p={6} backgroundColor='black' height='100vh' minHeight='500px'>
+                <Flex p={6} rounded='md' margin='auto' bgColor='white' direction='column' height='100%'>
+                    <Box mb={3} flex={0}>
+                        <Heading as='h2' size='xs' textColor='gray.500' mt={2} mb={0}>Address: </Heading>
+                        <Input name="address" bgColor='gray.300' value={address} onChange={(e) => handleInputChange(e.target.value)}></Input>
+                    </Box>
+                    <Flex flex={1} minHeight='300px' alignItems='stretch' gap={5}>
+                        <Box flexBasis='26em' flexGrow={0} overflow='auto'>
+                            <ConnectionsTable connections={connections} />
+                        </Box>
+                        <Box flexGrow={1} display={['none', 'none', 'flex']} p={5} rounded='md' backgroundColor='black' overflow='hidden' ref={graphRef}>
+                            <Box height='100%' width='100%' >
+                                <ConnectionsGraph connections={connections} width={width} height={height} />
+                            </Box>
+                        </Box>
+                    </Flex>
+                </Flex>
+            </Box>
+        </ChakraProvider>
+    )
+}
+
+export default ConnectionsPage;
